+++ wapp/__init__.py
@@ -1,0 +1,11 @@
+# -*- encoding: ascii -*-
+"""
+wapp namespace
+~~~~~~~~~~~~
+Whatsapp-analyzer libraries and apps.
+"""
+
+# WARNING: do not put any more code here.
+import pkgutil as _pkgutil
+
+__path__ = _pkgutil.extend_path(__path__, __name__)
