+++ wapp/analyzer/static/css/keen-dashboards.css
