+++ wapp/analyzer/static/css/styles.css
