+++ wapp/analyzer/static/js/stats.js
