+++ wapp/analyzer/templates/stats.html
